requirements.txt
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
GitPython>=3.1.40